(maint) Use explicit commands in the README
Prior to this commit, there was some bash shorthand in the readme. This
is prone to failues as customers may not run the commands exactly as
written. For example, if a customer ran 'ls' prior to 'cd !$', the
command would be incorrect. This commit updates those commands to be
explict.