(SUP-3054) README Updates
This commit adds how to use a file resource to distribute a copy of the
CA certificate to agents. Additionally, this commit reorganize the
usage steps for running the plan to distribute the new CA and removes
references to PCP transport because not all tasks/plans in ca_extend
support use via the Orchestrator.