(SUP-2742) README Updates
This commit adds an example of using --query with the upload plan.
Additionally, the instructions for using the upload plan have been
reorganized to display the differences and tradeoffs for using an
inventory file. Also, minor changes to phrasing and providing an
example of what users may set cert to for the upload plan.