(SUP-4347) Allow extend.sh to operate on Puppet 6 CAs
Prior to this commit, the `extend.sh` script would only work with
single-entry CAs laid down by Puppet 5 and earlier. Puppet 6 moved
to establishing a 2 entry CA that consists of a root and intermediate
certificate authority.

This commit updates the `extend.sh` script with logic for handling
two-entry Puppet 6 CAs. This commit does not add support for renewing
externally issued CA certificates or for CA bundles that have been modified
to contain additional certificates that are not part of the Puppet CA.